requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http
azure-monitor-opentelemetry